requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-auth==1.21.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
idna==2.10
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
